add configuration for 2nd monitor
The 2nd monitor is here assumed to be to the right of the main monitor.
lxpanel now has a 2nd panel on the 2nd monitor(*) and pcmanfm will now
show a 2nd desktop on the 2nd monitor.
lxpanel's taskbar will only show windows on the same monitor as the panel.
Openbox will open new windows on the monitor with the mouse cursor.
Openbox will always position Thunderbird on the first monitor though.

* Not sure why lxpanel calls this panel "left".