keyboard layout: add setxkbmap back
On the desktop (but not the laptop!), when setxkbmap is not used,
if the compose key is held and another key is pressed,
the compose version of that other key will not be produced
(but it will be produced if compose is tapped and then that
other key is tapped).

Now both setxkbmap and xmodmap are used, with xmodmap
redundantly setting right alt as the compose key since
xmodmap sometimes reverts the settings set by setxkbmap
(as it does on the laptop!).