vim: text: fix performance problems
e7b9230abfdeeabb1ab4a3de4268d6a68db403bc introduced a performance
regression that caused some plugins (like vimtex when repeatedly
exiting and then re-entering a quickfix window) to perform
extremely poorly (slowly) (particularly when the opening/closing
of buffers/windows was involved). This fixes the regression by
only resetting the filetype when it seems the situation that
requires it to be reset has occurred.