vim: ultisnips: tex: remove leading space on infix operators/relations
I usually type a space before typing out the snippet
activation sequence, which results in two spaces being placed
before the infix operation/relation. The better solution would
be to have the snippet engine detect if there is a preceding
space, and if there is not, only then place a space before the
infix operation/relation.