vim: enable truecolor support on any terminal that reports as a color xterm
Before this, we checked the $XTERM_VERSION environment variable
to make sure a terminal reporting as xterm really was (or likely was)
xterm. We now don't take steps to check if a $TERM value containing
"xterm" is actually xterm, which allows truecolor support to be
enabled on terminals like LXTerminal and the Windows Console
(at the small risk of enabling truecolor on a terminal reporting as
xterm that does not actually support truecolor).