tmux: paste twice (or thrice) on double (or triple) middle/right-click
Before this, if we were to double middle-click to paste from
primary-select twice, the paste action would only occur once
because the second click would not count as "MouseDown2Pane"
but as "SecondClick2Pane" (and "DoubleClick2Pane" if there was
no third click). Triple clicking would also result in pasting
only once. The above of course would also apply, for the same
reason, to pasting from the clipboard with a right-click.

Now pasting with middle-click and right-click is defined for
MouseDown#Pane, SecondClick#Pane, and TripleClick#Pane
(where '#' takes value 2 (for middle-click) or 3 (for right-click)),
enabling rapid-fire repeat pasting from the mouse.