xcompose: restructure math section
Instead of having a whole section just for n-ary operators,
we use the same bindings as the small versions of those
operators (when a small version exists), but with the
<Multi_key> prefixed. When the small version does not exists,
a completely new binding is made in the appropriate section.

Since we no longer have a section just for the n-ary operators,
we repurpose the "<Multi_key> + m + B" binding for double struck
(blackboard bold) symbols (instead of retaining the more cumbersome
"<Multi_key> + m + s + B" binding).