tmux: refactor hooks
Besides the use of the ternary instead of the more cumbersome if, `set`
is used without the `-g` flag in the last two lines. This prevents the
status bar's on/off status from unwantedly being universally shared
among sessions from different session groups (before this change, if
session A and B [not from the same session group] both had two windows
open, and session B closed one window causing the status bar on session
B to switch off, then the status bar would also switch off for session A
even though session A would still have more than one window open).
Unfortunately, this comes along with it's own new bug.

BUG:
If session A and B are both in the same session group and the session
group has two windows open, then closing a window on session B will
cause the status bar to be switched off on session B, but the status bar
on session A will still have the status bar on despite also now only
having one window open.

We accept the new bug over the old behavior since having a status bar on
when it should be off is a much better situation than having a status
bar off when it should be on.