vim: text: rework how text is set to the default filetype (very hacky)
Before this (admittedly hacky) commit, there were some weird edge cases
that resulted in weird behavior.

For example, if we issued `find ~` to open the home directory in netrw,
all the netrw key-bindings would be disabled (most noticeably the <cr>
binding) if ~/.vim/after/ftplugin/text.vim contained
`let b:undo_ftplugin .= '|mapclear <buffer> | mapclear! <buffer>'`.
Perhaps this is because when the buffer for netrw is first created,
it still hasn't been given a filetype yet, so the
`autocmd BufEnter * if &filetype ==# '' ...` goes into effect and
sets the filetype to text. Then finally the buffer filetype gets set
to "netrw" and the `b:undo_ftplugin` commands are issued before
the netrw ftplugin is loaded. However, netrw creates the bindings for
the netrw buffer not in the netrw ftplugin but instead in
$VIMRUNTIME/autoload/netrw.vim, which perhaps is loaded in the netrw
buffer before the filetype is set to netrw (and thus these bindings
are wiped out by the `b:undo_ftplugin` of the old text ftplugin when
the filetype changes from "text" to "netrw").

For another example, if we have the autocmd
`autocmd BufRead ~/.xinitrc setfiletype sh`, then sometimes
(I noticed this always happening in terminal `vim` but never
happening in `gvim` so it may be a race condition [or it may
be related to the fact that ~/.vim/gvimrc has the line
`filetype detect`]) the filetype of ~/.xinitrc would be set
to "text" instead of "sh".

Now with this commit, these issues are resolved.
The work-around to these issues is to check if the
`&filetype` setting is empty by issuing
`&filetype ==# ''` with feedkeys (actually we call a
function [with feedkeys] that does this). By the time the
buffer is ready to accept keypresses, it appears any
filetype detection should have already occurred (such as
via something like `autocmd BufRead ~/.xinitrc setfiletype sh`).
Thus we will only set the filetype to "text" if the buffer
really wouldn't have been set to any other filetype.

I don't know if the assumption that all standard (not hacky
stuff like this script) filetype detection will be completed
by the time the buffer accepts keypresses is always true, or
if that observation is just the result of a race condition
going a particular way.

It would be nice to find a better solution to set the default
filetype to text, but for now this is the best I have.