vim: text: abandon feedkeys() approach (now use a timer)
feedkeys() has too many pitfalls, and having keys sent to buffers
could result in things going very wrong (e.g. it could result in
accidental execution of commands from a vim terminal). Thus we are
abandoning that approach, and we now just detect if the filetype is
already set after delaying for 50ms (despite that this is obviously
race-condition dependent).