vim: tex: make save buffer to tempfile use `write`
For some reason, with `saveas`, there would be an extraneous empty
buffer in the buffer list after activating the mapping. With `write`,
this doesn't happen, but has the issue that the vimtex command
`VimtexCompile` will fail because it still acts like the buffer isn't
associated with a filename. `edit` makes vimtex register that the buffer
is now associated with a filename.