dunst: remove comments
When there is a new dunst release, often the behavior of certain
settings changes, which means the comments describing the behavior of
those settings are then out of date. Instead of changing the comments in
the configuration file after every dunst release, we are removing them
entirely and instead just placing a single comment at the top that
referencing how to access documentation and examples.